Move graph into its own component
